refactor: reversing the names of Plugin and PluginContainer
Seemed backwards.  The PluginContainer is now in the host and is what a plugin is mounted inside, and Plugin wraps the actual plugin in its own app.
