+++ src/plugins/Plugin.jsx
@@ -1,40 +1,60 @@
-import React from 'react';
-
-import PluginIframe from './PluginIframe';
-import PluginErrorBoundary from './PluginErrorBoundary';
-
+import React, { useEffect, useMemo, useState } from 'react';
+import PropTypes from 'prop-types';
 import {
-  IFRAME_PLUGIN,
-} from './data/constants';
-import { pluginConfigShape } from './data/shapes';
+  dispatchMountedEvent, dispatchReadyEvent, dispatchUnmountedEvent, useHostEvent,
+} from './data/hooks';
+import { PLUGIN_RESIZE } from './data/constants';
 
-export default function Plugin({ config, ...props }) {
-  if (config === null) {
-    return null;
-  }
+export default function Plugin({
+  children, className, style, ready,
+}) {
+  const [dimensions, setDimensions] = useState({
+    width: null,
+    height: null,
+  });
 
-  let renderer = null;
-  switch (config.type) {
-    case IFRAME_PLUGIN:
-      renderer = (
-        <PluginIframe config={config} {...props} />
-      );
-      break;
-    // istanbul ignore next: default isn't meaningful, just satisfying linter
-    default:
-  }
+  const finalStyle = useMemo(() => ({
+    ...dimensions,
+    ...style,
+  }), [dimensions, style]);
+
+  useHostEvent(PLUGIN_RESIZE, ({ payload }) => {
+    setDimensions({
+      width: payload.width,
+      height: payload.height,
+    });
+  });
+
+  useEffect(() => {
+    dispatchMountedEvent();
+
+    return () => {
+      dispatchUnmountedEvent();
+    };
+  }, []);
+
+  useEffect(() => {
+    if (ready) {
+      dispatchReadyEvent();
+    }
+  }, [ready]);
 
   return (
-    <PluginErrorBoundary>
-      {renderer}
-    </PluginErrorBoundary>
+    <div className={className} style={finalStyle}>
+      {children}
+    </div>
   );
 }
 
 Plugin.propTypes = {
-  config: pluginConfigShape,
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+  ready: PropTypes.bool,
+  style: PropTypes.object, // eslint-disable-line
 };
 
 Plugin.defaultProps = {
-  config: null,
+  className: null,
+  style: {},
+  ready: true,
 };
