+++ src/plugins/index.js
@@ -2,10 +2,10 @@ export {
   usePluginSlot,
 } from './data/hooks';
 export {
-  default as Plugin,
+  default as PluginContainer,
 } from './Plugin';
 export {
-  default as PluginContainer,
+  default as Plugin,
 } from './PluginContainer';
 export {
   default as PluginSlot,
